Move `dummy_centerline` into `test_centerline.py`
Notes:

- This one is a little trickier; since `dummy_centerline` is used in
  two places, we have to do a relative import from one test file to
  the other. I prefer this to duplicating the function, but,
  this is the one case where I'm wondering if it would be better
  to centralize the dummy data test fixtures into a single file.

- That would be a much larger effort, however, and it would require
  taking stock of every bit of dummy data we've spread across all
  of our tests, and seeing where there might be redundancy we can
  address.
