+++ testing/api/test_deepseg_sc.py
@@ -8,8 +8,9 @@ import nibabel as nib
 
 from spinalcordtoolbox.image import Image
 import spinalcordtoolbox.deepseg_.sc as deepseg_sc
-from spinalcordtoolbox.testing.create_test_data import dummy_centerline
 from spinalcordtoolbox.utils import sct_test_path
+
+from .test_centerline import dummy_centerline
 
 
 param_deepseg = [
